Inline the logic of the filter layout reader, allowing us to evolve this a little more
