+++ vortex-scan/src/filter.rs
@@ -1,0 +1,124 @@
+// SPDX-License-Identifier: Apache-2.0
+// SPDX-FileCopyrightText: Copyright the Vortex contributors
+
+use std::iter;
+
+use bit_vec::BitVec;
+use itertools::Itertools;
+use parking_lot::RwLock;
+use sketches_ddsketch::DDSketch;
+use vortex_error::{VortexExpect, vortex_err, vortex_panic};
+use vortex_expr::ExprRef;
+use vortex_expr::dynamic::DynamicExprUpdates;
+use vortex_expr::forms::cnf::cnf;
+
+/// The selectivity histogram quantile to use for reordering conjuncts. Where 0 == no rows match.
+const DEFAULT_SELECTIVITY_QUANTILE: f64 = 0.1;
+
+/// A [`FilterExpr`] splits boolean expressions into individual conjunctions, tracks
+/// statistics about selectivity, and uses this information to reorder the evaluation of the
+/// conjunctions in an attempt to minimize the work done.
+pub struct FilterExpr {
+    /// The conjuncts involved in the filter expression.
+    conjuncts: Vec<ExprRef>,
+    /// A histogram of the selectivity of each conjunct.
+    conjunct_selectivity: Vec<RwLock<DDSketch>>,
+    /// Dynamic expression trackers for each conjunct, incase they contain dynamic expressions.
+    dynamic_conjuncts: Vec<Option<DynamicExprUpdates>>,
+    /// The preferred ordering of conjuncts.
+    ordering: RwLock<Vec<usize>>,
+    /// The quantile to use from the selectivity histogram of each conjunct.
+    selectivity_quantile: f64,
+}
+
+impl FilterExpr {
+    pub(crate) fn new(expr: ExprRef) -> Self {
+        let conjuncts = cnf(expr);
+        let num_conjuncts = conjuncts.len();
+
+        let dynamic_conjuncts = conjuncts.iter().map(DynamicExprUpdates::new).collect_vec();
+
+        Self {
+            conjuncts,
+            conjunct_selectivity: iter::repeat_with(|| RwLock::new(DDSketch::default()))
+                .take(num_conjuncts)
+                .collect(),
+            dynamic_conjuncts,
+            // The initial ordering is naive, we could order this by how well we expect each
+            // comparison operator to perform. e.g. == might be more selective than <=? Not obvious.
+            ordering: RwLock::new((0..num_conjuncts).collect()),
+            selectivity_quantile: DEFAULT_SELECTIVITY_QUANTILE,
+        }
+    }
+
+    /// The conjuncts that make up this filter expression.
+    pub fn conjuncts(&self) -> &[ExprRef] {
+        &self.conjuncts
+    }
+
+    /// The dynamic updates for the given conjunct, if any.
+    pub fn dynamic_updates(&self, conjunct_idx: usize) -> Option<&DynamicExprUpdates> {
+        self.dynamic_conjuncts[conjunct_idx].as_ref()
+    }
+
+    /// Returns the next preferred conjunct to evaluate.
+    pub fn next_conjunct(&self, remaining: &BitVec) -> Option<usize> {
+        let read = self.ordering.read();
+        // Take the first remaining conjunct in the ordered list.
+        read.iter().find(|&idx| remaining[*idx]).copied()
+    }
+
+    /// Report the selectivity of a conjunct, i.e. 0 means no rows matched the predicate.
+    #[allow(clippy::cast_possible_truncation)]
+    pub fn report_selectivity(&self, conjunct_idx: usize, selectivity: f64) {
+        if !(0.0..=1.0).contains(&selectivity) {
+            vortex_panic!(
+                "selectivity {} must be in the range [0.0, 1.0]",
+                selectivity
+            );
+        }
+
+        {
+            let mut histogram = self.conjunct_selectivity[conjunct_idx].write();
+
+            histogram.add(selectivity);
+        }
+
+        let all_selectivity = self
+            .conjunct_selectivity
+            .iter()
+            .map(|histogram| {
+                histogram
+                    .read()
+                    .quantile(self.selectivity_quantile)
+                    .map_err(|e| vortex_err!("{e}")) // Only errors when the quantile is out of range
+                    .vortex_expect("quantile out of range")
+                    // If the sketch is empty, its selectivity is 0.
+                    .unwrap_or_default()
+            })
+            .collect::<Vec<_>>();
+
+        {
+            let ordering = self.ordering.read();
+            if ordering.is_sorted_by_key(|&idx| all_selectivity[idx]) {
+                return;
+            }
+        }
+
+        // Re-sort our conjuncts based on the new statistics.
+        let mut ordering = self.ordering.write();
+        ordering.sort_unstable_by(|&l_idx, &r_idx| {
+            all_selectivity[l_idx]
+                .partial_cmp(&all_selectivity[r_idx])
+                .vortex_expect("Can't compare selectivity values")
+        });
+
+        log::debug!(
+            "Reordered conjuncts based on new selectivity {:?}",
+            ordering
+                .iter()
+                .map(|&idx| format!("({}) => {}", self.conjuncts[idx], all_selectivity[idx]))
+                .join(", ")
+        );
+    }
+}
